Remove unnecessary mappingController. Move user-items call to submission controller
